review - move types and queries, fix invalidating auth/self after user updates
