+++ client/src/app/api/models.ts
@@ -1027,3 +1027,63 @@ export interface AnalysisProfile {
   scope: AnalysisProfileScope;
   rules: AnalysisProfileRules;
 }
+
+export interface UserManagementResource {
+  id: number;
+  createUser: string;
+  updateUser: string;
+  createTime: string; // ISO 8601 datetime
+}
+
+export interface User extends UserManagementResource {
+  subject: string;
+  login: string; // required, immutable after creation
+  name: string;
+  password: string; // required on create (max 72 chars), masked in responses
+  email: string; // required
+  roles: Ref[];
+  tokens: Ref[];
+}
+
+export interface UserRole extends UserManagementResource {
+  name: string; // required
+  permissions: Ref[];
+}
+
+/** Permission REST resource (read-only via API). */
+export interface Permission extends UserManagementResource {
+  name: string; // required
+  scope: string; // required, unique
+}
+
+export interface AuthMe {
+  user?: User;
+  scopes: string[];
+}
+
+export interface Token extends UserManagementResource {
+  kind: string; // "api-key" | "access"
+  authId?: string;
+  subject?: string;
+  scopes?: string[];
+  issued?: string; // ISO 8601
+  expiration?: string; // ISO 8601
+  lifespan?: number; // remaining hours until expiration
+  grant?: Ref | null;
+  task?: Ref | null;
+  user?: Ref | null;
+  idpIdentity?: Ref | null;
+  idpClient?: Ref | null;
+}
+
+/**
+ * PAT (Personal Access Token) — request body for POST,
+ * and the response from POST (which includes the one-time raw secret).
+ * Also aliased as APIKey.
+ */
+export interface PersonalAccessToken {
+  id?: number;
+  lifespan?: number; // requested lifespan in hours (default: ~87600)
+  expiration?: string; // ISO 8601, alternative to lifespan
+  token?: string; // raw secret, only present in the create response
+}
