+++ client/src/app/pages/user-management/roles/use-roles.tsx
@@ -1,71 +1,11 @@
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { AxiosError } from "axios";
 import { useTranslation } from "react-i18next";
 
-import { createRole, deleteRole, getRoles, updateRole } from "@app/api/rest";
 import { useNotifications } from "@app/components/NotificationsContext";
-
-import { Role } from "../types";
-
-export const RolesQueryKey = "roles";
-
-export const useFetchRoles = () => {
-  const { data, isLoading, error } = useQuery<Role[], AxiosError>({
-    queryKey: [RolesQueryKey],
-    queryFn: getRoles,
-  });
-  return { roles: data ?? [], isLoading, fetchError: error };
-};
-
-export const useCreateRoleMutation = (
-  onSuccess?: () => void,
-  onError?: () => void
-) => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: createRole,
-    onSuccess: () => {
-      queryClient
-        .invalidateQueries({ queryKey: [RolesQueryKey] })
-        .then(() => onSuccess?.());
-    },
-    onError: () => onError?.(),
-  });
-};
-
-export const useUpdateRoleMutation = (
-  onSuccess?: () => void,
-  onError?: () => void
-) => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: updateRole,
-    onSuccess: () => {
-      queryClient
-        .invalidateQueries({ queryKey: [RolesQueryKey] })
-        .then(() => onSuccess?.());
-      onSuccess?.();
-    },
-    onError: () => onError?.(),
-  });
-};
-
-export const useDeleteRoleMutation = (
-  onSuccess?: () => void,
-  onError?: () => void
-) => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: deleteRole,
-    onSuccess: () => {
-      queryClient
-        .invalidateQueries({ queryKey: [RolesQueryKey] })
-        .then(() => onSuccess?.());
-      onSuccess?.();
-    },
-    onError: () => onError?.(),
-  });
-};
+import {
+  useCreateRoleMutation,
+  useDeleteRoleMutation,
+  useUpdateRoleMutation,
+} from "@app/queries/roles";
 
 export const useRoleActionsWithNotifications = () => {
   const { t } = useTranslation();
